backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
pydantic==2.7.4
requests==2.31.0
python-multipart==0.0.9
python-dotenv==1.0.1
psycopg2-binary==2.9.9